+++ CHANGELOG.md
@@ -38,3 +38,4 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 - Update default CO Person Role entries without linking to a COU if not applicable
 - CO Person's email gets verified during the registration process
 - Add global scope for `Localization` variables of the default CO, COManage. This CO is only accessible by the platform administors.
+- Allow CO Person to view all Org Identities linked to his/her profile